+++ src/pages/BridgePage/SendButton.tsx
@@ -15,6 +15,7 @@ import { useMantaWallet } from 'contexts/mantaWalletContext';
 import { useKeyring } from 'contexts/keyringContext';
 import { useBridgeTx } from './BridgeContext/BridgeTxContext';
 import { useBridgeData } from './BridgeContext/BridgeDataContext';
+import EvmTransferButton from './EvmBridge/ApproveTransferButton';
 
 const ValidationButton = () => {
   const config = useConfig();
@@ -180,54 +181,6 @@ const SendButton = () => {
       )}>
       Submit
     </button>
-  );
-};
-
-const EvmTransferButton = () => {
-  const {
-    isApiInitialized,
-    isApiDisconnected,
-    senderAssetType,
-    senderAssetCurrentBalance,
-    senderAssetTargetBalance,
-    originChain,
-    originApi,
-    originChainIsEvm,
-    originXcmAdapter,
-    destinationChain,
-    destinationAddress,
-    maxInput,
-    minInput,
-    senderNativeAssetCurrentBalance,
-    originFee
-  } = useBridgeData();
-
-  const { ethAddress, provider } = useMetamask();
-
-  const onClick = () => {
-    // call metamask to approve
-    const amount = senderAssetTargetBalance.valueAtomicUnits.toString();
-    await provider.request({
-      method: 'eth_sendTransaction',
-      params: [
-        {
-          from: ethAddress,
-          to: '0xd9b0DDb3e3F3721Da5d0B20f96E0817769c2B46D'
-        }
-      ]
-    });
-  };
-  return (
-    <div>
-      <button
-        onClick={onClick}
-        className={classNames(
-          'bg-connect-wallet-button py-2 unselectable-text cursor-pointer',
-          'text-center text-white rounded-lg w-full'
-        )}>
-        Approve
-      </button>
-    </div>
   );
 };
